FIN: Finished the code for factorial.
Will need to update the ReadMe to complete this process

+++ Factorial-Functionality/src/Factorial.java
@@ -1,13 +1,13 @@
 public class Factorial {
-    protected int linearFactorial(int x){
-        int out = 1;
+    protected long linearFactorial(int x){
+        long out = 1;
         for(int i = 1; i <= x; i++){
             out *= i;
         }
         return out;
     }
 
-    protected int recursiveFactorial(int x){
+    protected long recursiveFactorial(long x){
         if(x == 0){
             return 1;
         }
